refactor: restructure AppCard component and styles, moving to dedicated files
