+++ src/components/components.module.css
@@ -1,0 +1,5 @@
+*,
+*::before,
+*::after {
+  box-sizing: border-box;
+}
