+++ src/index.ts
@@ -1,5 +1,8 @@
+// Import reset CSS from embla (optional)
+import 'embla/reset.css';
+
 // Import and register all components
-import './components/app-card.tsx';
+import './components/app-card/app-card.tsx';
 
 // Initialize the application
 console.log('Application initialized');
